# reduce ernie_large test time