### PR types
Performance optimization

### PR changes
Others

### Describe
更改统一的线程数量，会造成部分模型的耗时增加，因此只修改ernie模型的线程数。CI机器上测试下性能加速比。

test_analyzer_ernie_large耗时，本地 37.89 sec -> 33.19sec
test_analyzer_ernie耗时，本地 7.54 sec->6.64 sec

### CI 机器上耗时

- 修改前
Test  #190: test_analyzer_ernie ................................   Passed   21.24 sec
Test  #191: test_analyzer_ernie_large ...................................   Passed  126.68 sec

- 修改后
Test  #190: test_analyzer_ernie .........................................   Passed   17.62 sec
Test  #191: test_analyzer_ernie_large ..........................   Passed  101.09 sec